+++ ng/src/geometry/faces.rs
@@ -14,7 +14,7 @@ pub trait Faces {
     ///
     /// `tolerance` defines by how far this triangulation is allowed to deviate
     /// from the faces' actual dimensions.
-    fn triangles(&self, tolerance: f32) -> Triangles;
+    fn triangles(&self, tolerance: f64) -> Triangles;
 }
 
 /// The triangles that approximate a shape's faces
@@ -65,7 +65,7 @@ impl From<[Point; 3]> for Triangle {
 }
 
 impl Faces for fj::Shape {
-    fn triangles(&self, tolerance: f32) -> Triangles {
+    fn triangles(&self, tolerance: f64) -> Triangles {
         match self {
             Self::Shape2d(shape) => shape.triangles(tolerance),
             Self::Shape3d(shape) => shape.triangles(tolerance),
@@ -74,7 +74,7 @@ impl Faces for fj::Shape {
 }
 
 impl Faces for fj::Shape2d {
-    fn triangles(&self, tolerance: f32) -> Triangles {
+    fn triangles(&self, tolerance: f64) -> Triangles {
         match self {
             Self::Circle(shape) => shape.triangles(tolerance),
             Self::Difference(shape) => shape.triangles(tolerance),
@@ -84,7 +84,7 @@ impl Faces for fj::Shape2d {
 }
 
 impl Faces for fj::Shape3d {
-    fn triangles(&self, tolerance: f32) -> Triangles {
+    fn triangles(&self, tolerance: f64) -> Triangles {
         match self {
             Self::Sweep(shape) => shape.triangles(tolerance),
         }
@@ -92,7 +92,7 @@ impl Faces for fj::Shape3d {
 }
 
 impl Faces for fj::Circle {
-    fn triangles(&self, tolerance: f32) -> Triangles {
+    fn triangles(&self, tolerance: f64) -> Triangles {
         let mut triangles = Triangles::new();
         let mut segments = self.edge_segments(tolerance);
 
@@ -113,14 +113,14 @@ impl Faces for fj::Circle {
 }
 
 impl Faces for fj::Difference {
-    fn triangles(&self, _tolerance: f32) -> Triangles {
+    fn triangles(&self, _tolerance: f64) -> Triangles {
         // TASK: Implement.
         todo!()
     }
 }
 
 impl Faces for fj::Square {
-    fn triangles(&self, _: f32) -> Triangles {
+    fn triangles(&self, _: f64) -> Triangles {
         let mut triangles = Triangles::new();
 
         let v = self.vertices();
@@ -133,7 +133,7 @@ impl Faces for fj::Square {
 }
 
 impl Faces for fj::Sweep {
-    fn triangles(&self, tolerance: f32) -> Triangles {
+    fn triangles(&self, tolerance: f64) -> Triangles {
         let mut triangles = Triangles::new();
 
         let original_triangles = self.shape.triangles(tolerance);
